requirements.txt
apache-airflow==2.5.0
mysql-connector-python==8.0.27
pandas==1.4.2
requests==2.28.1
